refactor(components): update FileList
